+++ app/src/main/java/com/aconno/sensorics/ui/MainActivity.kt
@@ -270,6 +270,14 @@ class MainActivity : DaggerAppCompatActivity(), EasyPermissions.PermissionCallba
                 startSettingsActivity()
                 return true
             }
+            R.id.action_buy_beacons -> {
+                BuyBeaconsActivity.start(this)
+                return true
+            }
+            R.id.action_about -> {
+                AboutActivity.start(this)
+                return true
+            }
             else -> return super.onOptionsItemSelected(item)
         }
     }