+++ domain/src/test/java/com/aconno/sensorics/domain/interactor/convert/ReadingToInputUseCaseTest.kt
@@ -17,7 +17,7 @@ class ReadingToInputUseCaseTest {
 
     @Test
     fun execute_Test() {
-
+        //TODO: implement this
     }
 
 }