+++ domain/src/test/java/com/aconno/sensorics/domain/interactor/ifttt/RestPublishUseCasesTest.kt
@@ -1,10 +1,14 @@
 package com.aconno.sensorics.domain.interactor.ifttt
 
 import com.aconno.sensorics.domain.ifttt.GeneralRestPublish
+import com.aconno.sensorics.domain.ifttt.RestPublish
 import com.aconno.sensorics.domain.ifttt.publish.RestPublishRepository
+import com.aconno.sensorics.domain.interactor.ifttt.publish.AddPublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.publish.DeletePublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.publish.GetAllEnabledPublishUseCase
+import com.aconno.sensorics.domain.interactor.ifttt.publish.GetAllPublishUseCase
 import com.aconno.sensorics.domain.interactor.ifttt.restpublish.GetRestPublishByIdUseCase
 import org.junit.Before
-import org.junit.Ignore
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mock
@@ -15,10 +19,11 @@ import org.mockito.junit.MockitoJUnitRunner
 @RunWith(MockitoJUnitRunner::class)
 class RestPublishUseCasesTest {
 
-    //    private lateinit var addRestPublishUseCase: AddRestPublishUseCase
-//    private lateinit var deleteRestPublishUseCase: DeleteRestPublishUseCase
-//    private lateinit var getAllRestPublishUseCase: GetAllRestPublishUseCase
+    private lateinit var addRestPublishUseCase: AddPublishUseCase<RestPublish>
+    private lateinit var deleteRestPublishUseCase: DeletePublishUseCase<RestPublish>
+    private lateinit var getAllRestPublishUseCase: GetAllPublishUseCase<RestPublish>
     private lateinit var getRestPublishByIdUseCase: GetRestPublishByIdUseCase
+    private lateinit var getAllEnabledRestPublishUseCase : GetAllEnabledPublishUseCase<RestPublish>
 
     private val id: Long = 0
 
@@ -39,60 +44,55 @@ class RestPublishUseCasesTest {
 
     @Before
     fun setUp() {
-//        addRestPublishUseCase = AddRestPublishUseCase(
-//            mockRestPublishRepository
-//        )
-//        deleteRestPublishUseCase = DeleteRestPublishUseCase(
-//            mockRestPublishRepository
-//        )
-//        getAllEnabledRestPublishUseCase = GetAllEnabledRestPublishUseCase(
-//            mockRestPublishRepository
-//        )
-//        getAllRestPublishUseCase = GetAllRestPublishUseCase(
-//            mockRestPublishRepository
-//        )
-//        getRestPublishByIdUseCase = GetRestPublishByIdUseCase(
-//            mockRestPublishRepository
-//        )
+        addRestPublishUseCase = AddPublishUseCase(
+            mockRestPublishRepository
+        )
+        deleteRestPublishUseCase = DeletePublishUseCase(
+            mockRestPublishRepository
+        )
+        getAllEnabledRestPublishUseCase = GetAllEnabledPublishUseCase(
+            mockRestPublishRepository
+        )
+        getAllRestPublishUseCase = GetAllPublishUseCase(
+            mockRestPublishRepository
+        )
+        getRestPublishByIdUseCase = GetRestPublishByIdUseCase(
+            mockRestPublishRepository
+        )
     }
 
-    @Ignore("Failing test")
     @Test
     fun testAddRestPublishUseCase() {
-//        addRestPublishUseCase.execute(restPublish).blockingGet()
+        addRestPublishUseCase.execute(restPublish).blockingGet()
 
         verify(mockRestPublishRepository).addPublish(restPublish)
         verifyNoMoreInteractions(mockRestPublishRepository)
     }
 
-    @Ignore("Failing test")
     @Test
     fun testDeleteRestPublishUseCase() {
-//        deleteRestPublishUseCase.execute(restPublish).blockingGet()
+        deleteRestPublishUseCase.execute(restPublish).blockingGet()
 
         verify(mockRestPublishRepository).deletePublish(restPublish)
         verifyNoMoreInteractions(mockRestPublishRepository)
     }
 
-    @Ignore("Failing test")
     @Test
     fun testGetAllEnabledRestPublishUseCase() {
-//        getAllEnabledRestPublishUseCase.execute()
+        getAllEnabledRestPublishUseCase.execute()
 
         verify(mockRestPublishRepository).allEnabled
         verifyNoMoreInteractions(mockRestPublishRepository)
     }
 
-    @Ignore("Failing test")
     @Test
     fun testGetAllRestPublishUseCase() {
-//        getAllRestPublishUseCase.execute()
+        getAllRestPublishUseCase.execute()
 
         verify(mockRestPublishRepository).all
         verifyNoMoreInteractions(mockRestPublishRepository)
     }
 
-    @Ignore("Failing test")
     @Test
     fun testGetRestPublishByIdUseCase() {
         getRestPublishByIdUseCase.execute(id)